fixes for iPA for iTunes

+++ ios/BBCNewsFeedController.cs
@@ -6,8 +6,8 @@ using System.Net;
 using System.Xml;
 using System.Collections.Generic;
 
-using MonoTouch.Foundation;
-using MonoTouch.UIKit;
+using Foundation;
+using UIKit;
 
 namespace RCSI
 {
@@ -39,12 +39,12 @@ namespace RCSI
 			_controller = controller;
 		}
 
-		public override int NumberOfSections (UITableView tableView)
+		public override nint NumberOfSections (UITableView tableView)
 		{
 			return 1;
 		}
 
-		public override int RowsInSection (UITableView tableView, int section)
+		public override nint RowsInSection (UITableView tableView, nint section)
 		{
 			return _feedItemList.Count;
 		}
